Refactor MBR module: Improve structure and functionality
- Updated import paths for clarity in Utilities.rs.
- Enhanced MBR_type struct with additional methods for managing partitions, including adding, removing, and validating partitions.
- Implemented methods for reading and writing MBR data to devices.
- Added comprehensive tests for MBR functionality, including partition management and validation.
- Improved display formatting for MBR information.
- Ensured proper handling of MBR signatures and partition overlaps.
